Correção de sort que o varoto m1lgr4u achoou

+++ CaixeiroViajante.h
@@ -10,8 +10,8 @@
 
 struct LinhaArquivoCV{
     int idNo;
-    int coordx;
-    int coordy;
+    double coordx;
+    double coordy;
 };
 
 struct ExecutionParams{
